Fix code review issues: use size_t for loop iterators and update assertion message
Address Copilot code review feedback:
- Change loop iterators from int to size_t in PresolverStatsTest.cpp to match num_presolvers type
- Update assertion message from "non-zero size" to "positive size" in libpapilo.cpp

+++ test/libpapilo/PresolverStatsTest.cpp
@@ -122,7 +122,7 @@ TEST_CASE( "per-presolver-statistics-are-tracked-correctly",
    size_t total_transactions = 0;
    int total_applied = 0;
 
-   for( int i = 0; i < static_cast<int>( num_presolvers ); ++i )
+   for( size_t i = 0; i < num_presolvers; ++i )
    {
       const char* name =
           libpapilo_statistics_get_presolver_name( statistics, i );
@@ -218,7 +218,7 @@ TEST_CASE( "per-presolver-statistics-match-overall-statistics",
        libpapilo_statistics_get_num_presolvers( statistics );
    size_t sum_applied = 0;
 
-   for( int i = 0; i < static_cast<int>( num_presolvers ); ++i )
+   for( size_t i = 0; i < num_presolvers; ++i )
    {
       size_t napplied =
           libpapilo_statistics_get_presolver_napplied( statistics, i );
